Connect VIP status view to state
